A bit of brrr in model parsing [3/3]
I do not possess the brain power to rewrite this but this makes it a bit faster. Note it'd be better to avoid reading value by value since struct is better when reading larger chunks. No fancy table this time but quick testing of all 3 changes shaved off a minute on Practice Range (02:51 > 01:53) and a minute and a half on King's Row (07:18 > 05:41). The rest of the hogs are in this parser, instancing and Blender being slow with shader nodes.